// src/store/modules/dispatch/mutation-types.ts
export enum MutationTypes {
    /** 设置调度显示模式 */
    DISPATCH__SET_DIAPATCH_MODE = 'DISPATCH/SET_DIAPATCH_MODE',

    /** 设置调度窗口大小 */
    DISPATCH__SET_DIAPATCH_SIZE = 'DISPATCH/SET_DIAPATCH_SIZE',

    /** 设置调度方案列表 */
    DISPATCH__SET_DIAPATCH_SCHEME_LIST = 'DISPATCH/SET_DIAPATCH_SCHEME_LIST',

    /** 设置选中的调度方案详情 */
    DISPATCH__SET_CURRENT_DIAPATCH_SCHEME_DETAIL = 'DISPATCH/SET_CURRENT_DIAPATCH_SCHEME_DETAIL',

    /** 当前选中的调度方案详情Id */
    DISPATCH__SET_SHIP_DISPATCH_SCHEME_ID = 'DISPATCH/SET_SHIP_DISPATCH_SCHEME_ID',
    /** 当前选中的调度方案详情Name */
    DISPATCH__SET_SHIP_DISPATCH_SCHEME_NAME = 'DISPATCH/SET_SHIP_DISPATCH_SCHEME_NAME',
    /** 当前选中的调度方案详情Code */
    DISPATCH__SET_SHIP_DISPATCH_SCHEME_CODE = 'DISPATCH/SET_SHIP_DISPATCH_SCHEME_CODE',

    /** 设置Query.Category */
    DISPATCH__SET_DIAPATCH_SCHEME_QUERY_CATEGORY = 'DISPATCH/SET_DIAPATCH_SCHEME_QUERY_CATEGORY',
    /** 设置Query.Datetimes */
    DISPATCH__SET_DIAPATCH_SCHEME_QUERY_DATETIMES = 'DISPATCH/SET_DIAPATCH_SCHEME_QUERY_DATETIMES',

    /** 设置选中的调度方案详情-项 */
    DISPATCH__SET_CURRENT_DIAPATCH_SCHEME_DETAIL_ITEM = 'DISPATCH/SET_CURRENT_DIAPATCH_SCHEME_DETAIL_ITEM',
}